clean the code, debug mlflow for webapp
