+++ autorad/config/config.py
@@ -6,12 +6,9 @@ import sys
 import tempfile
 from pathlib import Path
 
-import sklearn
 from rich.logging import RichHandler
 
 import autorad
-
-# sklearn.set_config(transform_output="pandas")
 
 CONFIG_DIR = os.path.dirname(__file__)
 TEST_DATA_DIR = os.path.join(
@@ -31,8 +28,6 @@ else:
     RESULT_DIR = tempfile.mkdtemp()
 
 MODEL_REGISTRY = os.path.join(RESULT_DIR, "models")
-# os.makedirs(MODEL_REGISTRY, exist_ok=True)
-# mlflow.set_tracking_uri("file://" + MODEL_REGISTRY)
 
 PARAM_DIR = os.path.join(CONFIG_DIR, "pyradiomics_params")
 PRESETS = {
