+++ autorad/preprocessing/preprocessor.py
@@ -2,23 +2,19 @@ from __future__ import annotations
 
 import dataclasses
 import logging
+from pathlib import Path
 
+import joblib
 import pandas as pd
-import sklearn
 from imblearn.over_sampling import ADASYN, SMOTE, BorderlineSMOTE
 from sklearn.pipeline import Pipeline
-from sklearn.preprocessing import MinMaxScaler, StandardScaler
+from sklearn.preprocessing import StandardScaler
 
-from autorad.config import config
+from autorad.config import config, type_definitions
 from autorad.data.dataset import TrainingData, TrainingInput, TrainingLabels
-from autorad.feature_selection.selector import (
-    FailoverSelectorWrapper,
-    create_feature_selector,
-)
+from autorad.feature_selection.selector import create_feature_selector
 
 log = logging.getLogger(__name__)
-
-sklearn.set_config(transform_output="pandas")
 
 
 def get_not_none_kwargs(**kwargs):
@@ -130,12 +126,16 @@ class Preprocessor:
     def _build_pipeline(self):
         steps = []
         if self.standardize:
-            steps.append(("standardize", StandardScaler()))
+            steps.append(
+                (
+                    "standardize",
+                    StandardScaler().set_output(transform="pandas"),
+                )
+            )
         if self.feature_selection_method is not None:
             steps.append(
                 (
                     "select",
-                    # FailoverSelectorWrapper(
                     create_feature_selector(
                         method=self.feature_selection_method,
                         **get_not_none_kwargs(n_features=self.n_features),
@@ -157,6 +157,48 @@ class Preprocessor:
             )
         pipeline = Pipeline(steps)
         return pipeline
+
+
+def run_auto_preprocessing(
+    data: TrainingData,
+    result_dir: type_definitions.PathLike,
+    oversampling: bool = True,
+    feature_selection: bool = True,
+    selection_methods: list[str] | str = "all",
+):
+    if not feature_selection:
+        selection_setups = [None]
+    elif selection_methods == "all":
+        selection_setups = config.FEATURE_SELECTION_METHODS
+    else:
+        selection_setups = selection_methods
+
+    if oversampling:
+        oversampling_methods = config.OVERSAMPLING_METHODS
+    else:
+        oversampling_methods = [None]
+
+    preprocessed = {}
+    for selection_method in selection_setups:
+        preprocessed[selection_method] = {}
+        for oversampling_method in oversampling_methods:
+            preprocessor = Preprocessor(
+                standardize=True,
+                feature_selection_method=selection_method,
+                oversampling_method=oversampling_method,
+            )
+            try:
+                preprocessed[selection_method][
+                    oversampling_method
+                ] = preprocessor.fit_transform(data)
+            except AssertionError:
+                log.error(
+                    f"Preprocessing with {selection_method} and {oversampling_method} failed."
+                )
+        if not preprocessed[selection_method]:
+            del preprocessed[selection_method]
+    with open(Path(result_dir) / "preprocessed.pkl", "wb") as f:
+        joblib.dump(preprocessed, f)
 
 
 def create_oversampling_model(method: str, random_state: int = config.SEED):
